Mbot2: some small changes

+++ src/main/java/de/fhg/iais/roberta/connection/wired/mBot2/Mbot2Connector.java
@@ -67,6 +67,11 @@ public class Mbot2Connector extends AbstractConnector<Mbot2> {
                         LOG.info("Download user program");
                         try {
                             Pair<byte[], String> program = this.serverCommunicator.downloadProgram(this.brickData);
+                            if ( program.getFirst().length > 65534 ) {
+                                this.fire(State.ERROR_UPLOAD_TO_ROBOT.setAdditionalInfo("Program too large for robot. Must be less than 65534 bytes"));
+                                this.fire(State.WAIT_FOR_CMD);
+                                break;
+                            }
                             File tmp = File.createTempFile("mbot2_", program.getSecond());
                             tmp.deleteOnExit();
